fix: cast QFL to float32 to prevent FP16 dtype mismatch
_quality_focal_loss receives half-precision preds under AMP but
computes in float. The index assignment loss[pos, label] = value
fails when loss is Half and value is Float. Fix by casting pred to
float32 at the start of the function.

+++ yolox/models/dinox_head.py
@@ -470,6 +470,8 @@ class DINOXHead(nn.Module):
         Returns:
             Per-anchor loss ``(N,)`` summed over classes.
         """
+        # Compute in float32 for numerical stability (AMP sends half-precision)
+        pred = pred.float()
         pred_sigmoid = pred.sigmoid()
         # Step 1: background loss for all anchors, all classes
         zerolabel = pred.new_zeros(pred.shape)
@@ -482,7 +484,7 @@ class DINOXHead(nn.Module):
         if pos_mask.any():
             pos_inds = pos_mask.nonzero(as_tuple=False).squeeze(1)
             pos_labels = labels[pos_inds]
-            pos_scores = scores[pos_inds]
+            pos_scores = scores[pos_inds].float()
 
             scale = (pos_scores - pred_sigmoid[pos_inds, pos_labels]).abs().pow(beta)
             loss[pos_inds, pos_labels] = F.binary_cross_entropy_with_logits(
